+++ lambdas/es/indexer/document_queue.py
@@ -12,7 +12,11 @@ from elasticsearch import Elasticsearch, RequestsHttpConnection
 from elasticsearch.helpers import bulk
 
 from t4_lambda_shared.preview import ELASTIC_LIMIT_BYTES
-from t4_lambda_shared.utils import separated_env_to_iter
+from t4_lambda_shared.utils import (
+    PACKAGE_INDEX_SUFFIX,
+    get_quilt_logger,
+    separated_env_to_iter,
+)
 
 CONTENT_INDEX_EXTS = separated_env_to_iter("CONTENT_INDEX_EXTS") or {
     ".csv",
@@ -34,12 +38,9 @@ EVENT_PREFIX = {
 CHUNK_LIMIT_BYTES = int(os.getenv('CHUNK_LIMIT_BYTES') or 9_500_000)
 ELASTIC_TIMEOUT = 30
 MAX_BACKOFF = 360  # seconds
-MAX_RETRY = 4  # prevent long-running lambdas due to malformed calls
-# signifies that the object is truly deleted, not to be confused with
-# s3:ObjectRemoved:DeleteMarkerCreated, which we may see in versioned buckets
-# see https://docs.aws.amazon.com/AmazonS3/latest/dev/NotificationHowTo.html
+MAX_RETRY = 2  # prevent long-running lambdas due to malformed calls
 QUEUE_LIMIT_BYTES = 100_000_000  # 100MB
-RETRY_429 = 5
+RETRY_429 = 3
 
 
 # pylint: disable=super-init-not-called
@@ -71,6 +72,7 @@ class DocumentQueue:
             handle: str = '',
             metadata: str = '',
             pointer_file: str = '',
+            # this could the hash OR tag; to be used in _id primary key
             package_hash: str = '',
             package_stats: Dict[str, int] = None,
             tags: List[str] = (),
@@ -82,18 +84,24 @@ class DocumentQueue:
             comment: str = '',
             key: str,
             etag: str,
-            last_modified: datetime,
+            last_modified: str,
             size: int = 0
     ):
         """format event as a document and then queue the document"""
+        logger_ = get_quilt_logger()
         if not bucket or not key:
             raise ValueError(f"bucket={bucket} or key={key} required but missing")
+        is_delete_marker = False
         if event_type.startswith(EVENT_PREFIX["Created"]):
             _op_type = "index"
         elif event_type.startswith(EVENT_PREFIX["Removed"]):
             _op_type = "delete"
+            if event_type.endswith("DeleteMarkerCreated"):
+                is_delete_marker = True
+                # we index (not delete) delete markers to sync state with S3
+                _op_type = "index"
         else:
-            print("Skipping unrecognized event type {event_type}")
+            logger_.error("Skipping unrecognized event type %s", event_type)
             return
         # On types and fields, see
         # https://www.elastic.co/guide/en/elasticsearch/reference/master/mapping.html
@@ -102,24 +110,36 @@ class DocumentQueue:
         # can cause exceptions from ES
         index_name = bucket
         if doc_type == DocTypes.PACKAGE:
-            index_name += "_packages"
+            index_name += PACKAGE_INDEX_SUFFIX
         if not index_name:
             raise ValueError(f"Can't infer index name; bucket={bucket}, doc_type={doc_type}")
+        # core properties for all document types;
+        # see https://elasticsearch-py.readthedocs.io/en/6.3.1/helpers.html
         body = {
             "_index": index_name,
+            "_op_type": _op_type,  # determines if action is upsert (index) or delete
+            # TODO remove this; it's not meaningful since we use a different index
+            # type for object vs. package documents
             "_type": "_doc",
+            # TODO nest fields under "document" and maybe use _type:{package, object}
             "comment": comment,
             "etag": etag,
             "key": key,
-            "last_modified": last_modified.isoformat(),
+            "last_modified": last_modified,
             "size": size,
+            "delete_marker": is_delete_marker,
+            "version_id": version_id,
         }
         if doc_type == DocTypes.PACKAGE:
-            if not handle or not package_hash or not pointer_file:
-                raise ValueError("missing required argument for package document")
+            if not handle:
+                raise ValueError("missing required argument for package doc")
+            if _op_type == "index":
+                if not (pointer_file and package_hash):
+                    raise ValueError("missing required argument for package doc")
             if not (
                 package_stats is None
-                or isinstance(package_stats, dict) and {'total_files', 'total_bytes'}.issubset(package_stats)
+                or isinstance(package_stats, dict)
+                and {'total_files', 'total_bytes'}.issubset(package_stats)
             ):
                 raise ValueError("Malformed package_stats")
             body.update({
@@ -135,6 +155,10 @@ class DocumentQueue:
                     "package_stats": package_stats,
                 })
         elif doc_type == DocTypes.OBJECT:
+            if not version_id:
+                # ensure the same versionId and primary keys (_id) as given by
+                #  list-object-versions in the enterprise bulk_scanner
+                version_id = "null"
             body.update({
                 # Elastic native keys
                 "_id": f"{key}:{version_id}",
@@ -148,10 +172,9 @@ class DocumentQueue:
                 "meta_text": "",
                 "target": "",
                 "updated": datetime.utcnow().isoformat(),
-                "version_id": version_id
             })
         else:
-            print(f"Skipping unhandled document type: {doc_type}")
+            logger_.error("Skipping unexpected document type: %s", doc_type)
 
         self._append_document(body)
 
@@ -160,16 +183,16 @@ class DocumentQueue:
 
     def _append_document(self, doc):
         """append well-formed documents (used for retry or by append())"""
+        logger_ = get_quilt_logger()
         if doc.get("content"):
             # document text dominates memory footprint; OK to neglect the
             # small fixed size for the JSON metadata
             self.size += min(doc["size"], ELASTIC_LIMIT_BYTES)
+        logger_.debug("Appending document %s", doc)
         self.queue.append(doc)
 
-    def send_all(self):
-        """flush self.queue in 1-2 bulk calls"""
-        if not self.queue:
-            return
+    def _make_elastic(self):
+        """create elasticsearch client"""
         elastic_host = os.environ["ES_HOST"]
         session = boto3.session.Session()
         credentials = session.get_credentials().get_frozen_credentials()
@@ -183,7 +206,7 @@ class DocumentQueue:
             aws_service="es"
         )
 
-        elastic = Elasticsearch(
+        return Elasticsearch(
             hosts=[{"host": elastic_host, "port": 443}],
             http_auth=awsauth,
             max_backoff=get_time_remaining(self.context) if self.context else MAX_BACKOFF,
@@ -193,10 +216,58 @@ class DocumentQueue:
             verify_certs=True,
             connection_class=RequestsHttpConnection
         )
+
+    def _filter_and_delete_packages(self, elastic):
+        """handle package hard delete"""
+        logger_ = get_quilt_logger()
+        true_docs = []
+        for doc in self.queue:
+            pointer_file = doc.get("pointer_file")
+            # handle hard package delete outside of the bulk operation
+            if doc["_op_type"] == "delete" and pointer_file and not doc.get("delete_marker"):
+                index = doc.get("_index")
+                assert index.endswith(PACKAGE_INDEX_SUFFIX), f"Refuse to delete non-package: {doc}"
+                handle = doc.get("handle")
+                assert handle, "Cannot delete package without handle"
+                # no try/except because failure to delete, or trying to delete things
+                # that aren't in ES, does not throw
+                deletes = elastic.delete_by_query(
+                    index=index,
+                    body={
+                        "query": {
+                            "bool": {
+                                "must": [
+                                    # use match (not term) because some of these fields are analyzed
+                                    {"match": {"handle": handle}},
+                                    {"match": {"pointer_file": pointer_file}},
+                                    {"match": {"delete_marker": False}},
+                                ]
+                            }
+                        }
+                    },
+                    # we delete synchronously, so don't let it linger too long
+                    timeout='20s'
+                )
+                logger_.debug("Deleted %s stamped %s: %s", handle, pointer_file, deletes)
+                if not deletes.get("deleted"):
+                    logger_.warning("Unable to delete: %s", doc)
+            # send everything else to bulk()
+            else:
+                logger_.debug("Not filtering docs: %s", doc)
+                true_docs.append(doc)
+        # the queue is now everything we didn't delete by query above
+        self.queue = true_docs
+
+    def send_all(self):
+        """flush self.queue in 1-2 bulk calls"""
+        if not self.queue:
+            return
+        elastic = self._make_elastic()
         # For response format see
         # https://www.elastic.co/guide/en/elasticsearch/reference/6.7/docs-bulk.html
         # (We currently use Elastic 6.7 per quiltdata/deployment search.py)
         # note that `elasticsearch` post-processes this response
+        self._filter_and_delete_packages(elastic)
         _, errors = bulk_send(elastic, self.queue)
         if errors:
             id_to_doc = {d["_id"]: d for d in self.queue}
@@ -208,6 +279,9 @@ class DocumentQueue:
                         inner = error["index"]
                     if "delete" in error:
                         inner = error["delete"]
+                        # don't retry deleting things that aren't there
+                        if "not_found" in inner.get("result", ""):
+                            continue
                     if "_id" in inner:
                         doc = id_to_doc[inner["_id"]]
                         # Always retry the source document if we can identify it.
@@ -235,11 +309,11 @@ class DocumentQueue:
 
 def get_time_remaining(context):
     """returns time remaining in seconds before lambda context is shut down"""
+    logger_ = get_quilt_logger()
     time_remaining = floor(context.get_remaining_time_in_millis()/1000)
     if time_remaining < 30:
-        print(
-            f"Warning: Lambda function has less than {time_remaining} seconds."
-            " Consider reducing bulk batch size."
+        logger_.warning(
+            "Lambda function has {time_remaining} sec remaining. Reduce batch size?"
         )
 
     return time_remaining
@@ -247,6 +321,8 @@ def get_time_remaining(context):
 
 def bulk_send(elastic, list_):
     """make a bulk() call to elastic"""
+    logger_ = get_quilt_logger()
+    logger_.debug("bulk_send(): %s", list_)
     return bulk(
         elastic,
         list_,